user_registration/requirements.txt
Django>=4.2.7, <4.3
psycopg2-binary>=2.9.9, <3.10
djangorestframework>=3.14.0, <3.15
django-cors-headers>=4.3.0, <4.4